Move tests from DataSourceV2SQLSuite
